refactor(observability): 按领域边界拆分类型契约
