+++ src/observability/capture-coverage.ts
@@ -1,4 +1,4 @@
-import type { ObservationCaptureCoverage } from '../types/index.js';
+import type { ObservationCaptureCoverage } from './contracts/inbox.js';
 
 export function buildExplicitObservationCaptureCoverage(
   hasSubmittedEvidence: boolean,
