+++ src/observability/contracts/experience.ts
@@ -1,448 +1,6 @@
-/**
- * Observability DTO 集中定义。
- *
- * 解开 `src/observability/` 内部 4 个文件(experience / inbox / skill-chain /
- * experience-frontmatter)之间的 type-only 循环依赖：types/ 是叶子层，不依赖
- * 任何业务模块；observability/ 的运行时文件向这里 import 类型即可。
- *
- * 设计原则：
- *   - 只放纯 DTO(string union / interface / 类型别名)，不放任何运行时函数 / 常量
- *   - 不依赖 src 内的业务模块(observability / cli / server / renderer / diagnosis)
- *   - 跨层共享类型集中在 types/，单一来源
- */
-
-import type { DiagnosisBundle } from './diagnosis.js';
-import type { ToolCallStatus } from './executor.js';
-import type { TraceSourceKind } from './trace.js';
-import type { SkillHardRule, SkillWorkflow } from '../shared/hard-rules.js';
-
-// ---------- trace-source ----------
-
-export interface TraceIngestionSummary {
-  fileCount: number;
-  sourceRecordCount: number;
-  parsedRecordCount: number;
-  malformedRecordCount: number;
-  ignoredValueCount: number;
-  unknownEventCount: number;
-  filteredSessionCount: number;
-}
-
-export interface TraceSourceMetadata {
-  channel?: string;
-  sender?: string;
-  senderId?: string;
-  provider?: string;
-  model?: string;
-  modelApi?: string;
-  businessActions?: string[];
-}
-
-// ---------- review-state ----------
-
-export type ObservationReviewTargetType =
-  | 'experience_session'
-  | 'inbox_item'
-  | 'skill'
-  | 'goal_slice_correction'
-  | 'evidence_metric'
-  | 'reviewer_judgment'
-  | 'soft_standard'
-  | 'goal_keyword_correction'
-  | 'result_artifact_correction'
-  | 'completion_result_correction'
-  | 'deliverable_artifact_correction'
-  | 'skill_relevance_correction'
-  | 'workflow_completion_correction'
-  | 'hardrule_execution_correction'
-  | 'main_tool_execution_correction';
-export type ObservationReviewVerdict = 'reviewed' | 'real_issue' | 'not_issue' | 'needs_more_context' | 'confirmed' | 'rejected';
-export type ObservationMetricKey =
-  | 'user_correction'
-  | 'user_interruption'
-  | 'user_follow_up'
-  | 'negative_feedback'
-  | 'positive_feedback'
-  | 'hard_rule'
-  | 'user_goal_shift'
-  | 'result_artifact'
-  | 'completion_result'
-  | 'deliverable_artifact'
-  | 'progress_update'
-  | 'self_correction'
-  | 'repeated_execution';
-export type ObservationMetricScope = 'message' | 'skill_segment';
-
-export interface ObservationReviewStateEntry {
-  targetType: ObservationReviewTargetType;
-  targetId: string;
-  verdict: ObservationReviewVerdict;
-  reviewedAt: string;
-  note?: string;
-  reason?: string;
-  metricKey?: ObservationMetricKey;
-  metricScope?: ObservationMetricScope;
-  metricScopeId?: string;
-  traceId?: string;
-  sourceTrace?: string;
-  sessionId?: string;
-  messageIndex?: number;
-  messageUuid?: string;
-  callInstanceId?: string;
-  toolUseId?: string;
-  snippet?: string;
-}
-
-export interface ObservationReviewState {
-  kind: 'observe-review-state';
-  schemaVersion: 2;
-  updatedAt: string;
-  entries: Record<string, ObservationReviewStateEntry>;
-}
-
-export interface ObservationReviewStateUpdate {
-  targetType: ObservationReviewTargetType;
-  targetId: string;
-  verdict: ObservationReviewVerdict;
-  note?: string;
-  reason?: string;
-  metricKey?: ObservationMetricKey;
-  metricScope?: ObservationMetricScope;
-  metricScopeId?: string;
-  traceId?: string;
-  sourceTrace?: string;
-  sessionId?: string;
-  messageIndex?: number;
-  messageUuid?: string;
-  callInstanceId?: string;
-  toolUseId?: string;
-  snippet?: string;
-}
-
-// ---------- problem-patterns ----------
-
-export type ExperienceProblemBucket =
-  | 'output_format'
-  | 'content_accuracy'
-  | 'missing_context'
-  | 'rule_violation'
-  | 'workflow_mismatch'
-  | 'tool_runtime'
-  | 'goal_shift'
-  | 'unclear';
-
-export type ExperienceProblemSignal =
-  | 'user_correction'
-  | 'negative_feedback'
-  | 'user_interruption'
-  | 'hard_rule'
-  | 'user_goal_shift'
-  | 'tool_failure'
-  | 'workflow_mismatch'
-  | 'artifact_missing'
-  | 'observer_lifecycle_failed'
-  | 'orchestration_boundary_violation';
-
-export interface ExperienceProblemEvidenceRef {
-  id: string;
-  kind: string;
-  traceId?: string;
-  sourceTrace: string;
-  sessionId: string;
-  messageIndex?: number;
-  logicalMessageIndex?: number;
-  sourceLineIndex?: number;
-  messageUuid?: string;
-  callInstanceId?: string;
-  toolUseId?: string;
-  timestamp?: string;
-  role?: 'user' | 'assistant' | 'tool' | 'other';
-  label?: string;
-  snippet?: string;
-}
-
-export interface ExperienceProblemPattern {
-  id: string;
-  bucket: ExperienceProblemBucket;
-  patternKey: string;
-  count: number;
-  sessionCount: number;
-  recentSessionIds: string[];
-  signalTypes: ExperienceProblemSignal[];
-  evidenceRefs: ExperienceProblemEvidenceRef[];
-  lastSeen?: string;
-}
-
-export interface ProblemTimelineEvent {
-  id: string;
-  kind: string;
-  traceId?: string;
-  sourceTrace: string;
-  sessionId: string;
-  messageIndex?: number;
-  logicalMessageIndex?: number;
-  sourceLineIndex?: number;
-  messageUuid?: string;
-  callInstanceId?: string;
-  toolUseId?: string;
-  timestamp?: string;
-  role?: 'user' | 'assistant' | 'tool' | 'other';
-  label?: string;
-  snippet?: string;
-  fullText?: string;
-  toolName?: string;
-  isError?: boolean;
-}
-
-// ---------- skill-chain-advisories ----------
-
-export type SkillChainAdvisoryCode =
-  | 'hardrules_not_declared'
-  | 'workflows_not_declared'
-  | 'skill_md_not_found';
-
-export interface SkillChainAdvisory {
-  code: SkillChainAdvisoryCode;
-  message: string;
-  exampleYaml?: string;
-  commandTemplate?: string;
-  shortLabel: string;
-}
-
-// ---------- inbox ----------
-
-export type ObservationSignalType = 'failed_search' | 'repeated_failure' | 'hedging' | 'explicit_marker' | 'user_feedback';
-/** @deprecated Prefer TraceSourceKind for new source-neutral APIs. */
-export type ObservationSourceKind = TraceSourceKind;
-export type ObservationSeverityReasonCode =
-  | 'knowledge_gap_suspected'
-  | 'repeated_failure_suspected'
-  | 'explicit_gap_marker'
-  | 'exploratory_probe'
-  | 'skill_asset_unavailable'
-  | 'soft_hedging_signal'
-  | 'user_reported_knowledge_issue'
-  | 'tool_or_runtime_noise';
-export type ObservationSignalSubtype =
-  | 'hard_miss'
-  | 'repeated_failure'
-  | 'exploratory_miss'
-  | 'tool_error'
-  | 'permission_error'
-  | 'bash_probe'
-  | 'not_found'
-  | 'transient_file_missing'
-  | 'skill_asset_read_failed'
-  | 'permission_denied'
-  | 'tool_limit'
-  | 'tool_failure'
-  | 'regex_only'
-  | 'llm_classified'
-  | 'marker'
-  | 'explicit_user_feedback';
-
-export type ObservationCaptureObservedEventKind =
-  | 'tool_boundary'
-  | 'user_feedback'
-  | 'submitted_evidence';
-
-export type ObservationCaptureUnavailableEventKind =
-  | 'full_conversation'
-  | 'external_tool_calls'
-  | 'hidden_reasoning';
-
-export interface ObservationCaptureCoverage {
-  coverageStatus: 'partial';
-  capturePath: 'explicit_tool_call';
-  observedEventKinds: ObservationCaptureObservedEventKind[];
-  unavailableEventKinds: ObservationCaptureUnavailableEventKind[];
-}
-
-export interface ObservationEvidence {
-  traceId?: string;
-  sessionId?: string;
-  sourceTrace?: string;
-  sourceKind?: ObservationSourceKind;
-  tool?: string;
-  query?: string;
-  path?: string;
-  outputSnippet?: string;
-  assistantSnippet?: string;
-  userFeedbackSnippet?: string;
-  submittedEvidenceSnippet?: string;
-  markerToken?: string;
-  messageIndex?: number;
-  messageUuid?: string;
-  callInstanceId?: string;
-  toolUseId?: string;
-  segmentTimestamp?: string;
-}
-
-export interface ObservationMessageRef {
-  role: 'user' | 'assistant' | 'other';
-  snippet: string;
-  messageIndex: number;
-  uuid?: string;
-  timestamp?: string;
-}
-
-export interface ObservationMessageWindow {
-  before: ObservationMessageRef[];
-  event: ObservationMessageRef[];
-  after: ObservationMessageRef[];
-  resolutionAfter: 'resolved' | 'unresolved' | 'unknown';
-}
-
-export interface ObservationInboxItem {
-  id: string;
-  skillName: string;
-  artifactVersion: string | 'unknown';
-  artifactHash?: string;
-  cwd?: string;
-  sessionId: string;
-  /** Physical evidence stream identity; unlike sessionId, unique across reused run ids. */
-  traceId?: string;
-  sourceTrace: string;
-  sourceKind: ObservationSourceKind;
-  signalType: ObservationSignalType;
-  signalSubtype: ObservationSignalSubtype;
-  confidence: number;
-  attributionConfidence: number;
-  severity: 'high' | 'medium' | 'low' | 'noise';
-  severityReasonCode?: ObservationSeverityReasonCode;
-  severityReason?: string;
-  captureCoverage?: ObservationCaptureCoverage;
-  evidence: ObservationEvidence;
-  messageWindow?: ObservationMessageWindow;
-  firstSeen: string;
-  lastSeen: string;
-  occurrences: number;
-  /** Occurrences backed by an observed source timestamp. */
-  timestampedOccurrences?: number;
-  recentSessionIds: string[];
-  recentTraceIds?: string[];
-  representativeEvidence: ObservationEvidence[];
-}
-
-export interface ObservationSessionTimeRange {
-  sessionId: string;
-  traceId?: string;
-  sessionGroupId?: string;
-  sourceTrace: string;
-  sourceKind: ObservationSourceKind;
-  traceRole?: 'standalone' | 'main' | 'subagent';
-  traceLabel?: string;
-  cwd?: string;
-  startTimestamp?: string;
-  endTimestamp?: string;
-  durationMs?: number;
-}
-
-export type ObservationSourceRecordArchiveStatus = 'available' | 'partial' | 'unavailable';
-export type ObservationSourceRecordArchiveReason =
-  | 'no_record_ranges'
-  | 'source_missing'
-  | 'unsupported_source'
-  | 'read_failed'
-  | 'archive_limit';
-
-/**
- * A report only retains a relative pointer to its bounded source-record archive.
- * The archive itself stays beside the report so Studio never follows an
- * arbitrary absolute path supplied by report JSON.
- */
-export interface ObservationSourceRecordArchiveRef {
-  experienceSessionId: string;
-  status: ObservationSourceRecordArchiveStatus;
-  relativePath?: string;
-  recordCount: number;
-  omittedRecordCount: number;
-  byteCount: number;
-  truncated: boolean;
-  reason?: ObservationSourceRecordArchiveReason;
-}
-
-export interface ObservationSourceRecord {
-  sourceIndex: number;
-  traceId: string;
-  sourceTrace: string;
-  sourceType: string;
-  sourceEventId?: string;
-  timestamp?: string;
-  raw: string;
-  byteCount: number;
-  truncated: boolean;
-  redacted: boolean;
-}
-
-export interface ObservationSourceRecordArchive {
-  archiveKind: 'observe-source-records';
-  schemaVersion: 1;
-  experienceSessionId: string;
-  generatedAt: string;
-  records: ObservationSourceRecord[];
-  omittedRecordCount: number;
-  byteCount: number;
-  truncated: boolean;
-}
-
-export interface ObservationSourceRecordArchiveView {
-  status: ObservationSourceRecordArchiveStatus;
-  /** Total retained records, even when `records` is intentionally lazy. */
-  recordCount: number;
-  records: ObservationSourceRecord[];
-  omittedRecordCount: number;
-  byteCount: number;
-  truncated: boolean;
-  reason?: ObservationSourceRecordArchiveReason | 'archive_invalid';
-}
-
-export interface ObservationInboxReport {
-  kind: 'observe-inbox';
-  schemaVersion: 2;
-  meta: {
-    tracePath: string;
-    generatedAt: string;
-    sessionCount?: number;
-    sessionTimeRange?: {
-      from: string;
-      to: string;
-      durationMs?: number;
-    };
-    sessionTimeRanges?: ObservationSessionTimeRange[];
-    ingestion?: TraceIngestionSummary;
-    segmentCount: number;
-    itemCount: number;
-    skillInvocationCounts?: Record<string, number>;
-    skillSessionCounts?: Record<string, number>;
-    skillInvocationLastSeen?: Record<string, string>;
-    skillToolCallCounts?: Record<string, Record<string, number>>;
-    timestampedSegmentCount?: number;
-    timestampCoverage?: number;
-    sourceRecordArchives?: ObservationSourceRecordArchiveRef[];
-  };
-  items: ObservationInboxItem[];
-  experience?: ObservationExperienceReport;
-  diagnostics?: DiagnosisBundle;
-}
-
-export interface ObservationSkillRollup {
-  skillName: string;
-  invocationCount: number;
-  sessionCount: number;
-  observationCount: number;
-  highCount: number;
-  mediumCount: number;
-  lowCount: number;
-  noiseCount: number;
-  latestSeen: string;
-}
-
-export interface BuildObservationInboxReportOptions {
-  reviewState?: ObservationReviewState;
-}
-
-// ---------- experience: string unions ----------
+import type { ToolCallStatus } from '../../types/executor.js';
+import type { ExperienceProblemPattern } from './problem-patterns.js';
+import type { ObservationSourceKind, TraceSourceMetadata } from './trace.js';
 
 export type ExperienceReviewPriority = 'review_first' | 'sample_review' | 'routine_sample';
 export type ExperienceGoalSliceReasonCode = 'skill_segment_boundary' | 'explicit_user_goal_shift' | 'default_session_slice';
@@ -572,31 +130,6 @@ export interface ExperienceTimelineEvent extends ExperienceEvidenceRef {
 
 // ---------- Knowledge Debugger task trajectory ----------
 
-export type DebugKnowledgeKind = 'project_instruction' | 'skill' | 'runtime_evidence';
-export type DebugKnowledgeAccessKind = 'injected' | 'read' | 'returned';
-export type TaskReplayStepKind =
-  | 'user_request'
-  | 'user_message'
-  | 'user_correction'
-  | 'runtime_context'
-  | 'skill_context'
-  | 'tool_exchange'
-  | 'unmatched_tool_result'
-  | 'assistant_message'
-  | 'model_activity'
-  | 'lifecycle'
-  | 'observation'
-  | 'system_event';
-export type TaskReplayIntegrityCode =
-  | 'task_boundary_unavailable'
-  | 'timeline_truncated'
-  | 'malformed_records'
-  | 'ignored_values'
-  | 'unknown_events'
-  | 'unmatched_tool_calls'
-  | 'unmatched_tool_results'
-  | 'missing_timestamps';
-
 export type TaskWindowBasis =
   | 'turn_id'
   | 'turn_lifecycle'
@@ -631,144 +164,6 @@ export interface ExperienceTurnSummary {
   assistantMessageCount: number;
   toolCallCount: number;
   toolFailureCount: number;
-}
-
-export interface ConversationTaskItem {
-  turnId: string;
-  sourceTurnId?: string;
-  /** Existing observe reports retain their report-scoped session route. */
-  experienceSessionId?: string;
-  /** Preferred source-neutral task trajectory route. */
-  trajectoryHref?: string;
-  title: string;
-  startTimestamp?: string;
-  endTimestamp?: string;
-  durationMs?: number;
-  status: ExperienceTurnStatus;
-  eventCount: number;
-  toolCallCount: number;
-  toolFailureCount: number;
-  relatedSkillNames: string[];
-}
-
-export interface ConversationListItem {
-  threadId: string;
-  sourceThreadId: string;
-  sourceKind: ObservationSourceKind;
-  title: string;
-  preview?: string;
-  cwd?: string;
-  model?: string;
-  reasoningEffort?: string;
-  archived?: boolean;
-  tokensUsed?: number;
-  childThreadCount?: number;
-  startTimestamp?: string;
-  endTimestamp?: string;
-  durationMs?: number;
-  /** Undefined until the selected rollout has been indexed. */
-  turnCount?: number;
-  toolCallCount?: number;
-  toolFailureCount?: number;
-  relatedSkillNames: string[];
-  tasks: ConversationTaskItem[];
-}
-
-export interface ConversationIndexViewModel {
-  conversations: ConversationListItem[];
-  totalTurnCount: number;
-  totalToolCallCount: number;
-  totalToolFailureCount: number;
-  indexedConversationCount?: number;
-  unarchivedConversationCount?: number;
-  archivedConversationCount?: number;
-  workspaceCount?: number;
-}
-
-/**
- * Source-neutral task trajectory input. Observe reports carry more review and
- * Skill-attribution fields, but the debugger only requires this trace slice.
- */
-export interface TaskTrajectorySession {
-  id: string;
-  threadId: string;
-  sourceThreadId: string;
-  sessionId: string;
-  sourceTrace: string;
-  sourceKind: ObservationSourceKind;
-  entrypoint?: string;
-  sourceMetadata?: TraceSourceMetadata;
-  cwd?: string;
-  startTimestamp?: string;
-  endTimestamp?: string;
-  attributedEventIds: string[];
-  turns: ExperienceTurnSummary[];
-  fullSessionTimeline: ExperienceTimelineEvent[];
-  indicators?: Pick<ExperienceReviewIndicators, 'userCorrectionCount'>;
-  sessionStory?: ExperienceSessionStory;
-}
-
-export interface TaskWindowScope {
-  basis: TaskWindowBasis;
-  turnId?: string;
-  normalizedEventCount: number;
-  semanticEventCount: number;
-  attributedEventCount: number;
-  matchedAttributedEventCount: number;
-  truncated: boolean;
-}
-
-export interface DebugKnowledgeEvidence {
-  id: string;
-  knowledgeKind: DebugKnowledgeKind;
-  accessKind: DebugKnowledgeAccessKind;
-  label: string;
-  sourceLocator?: string;
-  contentHash?: string;
-  firstSeen?: string;
-  lastSeen?: string;
-  accessCount: number;
-  evidenceRefs: ExperienceEvidenceRef[];
-}
-
-export interface TaskReplayStep {
-  id: string;
-  order: number;
-  stepKind: TaskReplayStepKind;
-  timestamp?: string;
-  title: string;
-  events: ExperienceTimelineEvent[];
-  toolStatus?: ToolCallStatus;
-  knowledgeEvidenceIds: string[];
-}
-
-export interface TaskReplayIntegrityNotice {
-  code: TaskReplayIntegrityCode;
-  count: number;
-}
-
-export interface KnowledgeDebuggerViewModel {
-  session: TaskTrajectorySession;
-  taskScope: TaskWindowScope;
-  summary: {
-    userGoal?: string;
-    finalResponse?: string;
-    observedStartTimestamp?: string;
-    observedEndTimestamp?: string;
-    toolCallCount: number;
-    toolFailureCount: number;
-    hasUserCorrection: boolean;
-    /** Distinct models observed inside the attributed task window, in first-seen order. */
-    observedModels: string[];
-  };
-  steps: TaskReplayStep[];
-  normalizedEvents: ExperienceTimelineEvent[];
-  sourceRecords: ObservationSourceRecordArchiveView;
-  knowledgeEvidence: DebugKnowledgeEvidence[];
-  integrity: {
-    status: 'complete' | 'partial';
-    notices: TaskReplayIntegrityNotice[];
-  };
 }
 
 export interface ExperienceTimelineBranch {
@@ -1359,123 +754,4 @@ export interface ObservationExperienceReport {
   invocations: ExperienceInvocation[];
   sessions: ExperienceSessionSummary[];
   skills: ExperienceSkillSummary[];
-}
-
-// ---------- skill-chain ----------
-
-export type ObservationRuntimeCheckStatus = 'passed' | 'attention' | 'manual_review';
-
-export interface ObservationRuntimeCheck {
-  nodeKind: 'hardRule' | 'workflowNode';
-  id: string;
-  title: string;
-  expectation: string;
-  status: ObservationRuntimeCheckStatus;
-  reason: string;
-  evidenceCount: number;
-  evidenceSnippets: string[];
-}
-
-export type SkillRuntimeEvidencePackSourceType =
-  | 'tool_call'
-  | 'tool_result'
-  | 'assistant_message'
-  | 'user_feedback'
-  | 'artifact'
-  | 'runtime_context'
-  | 'skill_context'
-  | 'unknown';
-
-export interface SkillRuntimeEvidencePackRef extends Pick<ExperienceEvidenceRef,
-  'id' | 'kind' | 'sourceTrace' | 'sessionId' | 'messageUuid' | 'messageIndex' | 'logicalMessageIndex' | 'sourceLineIndex' | 'callInstanceId' | 'toolUseId' | 'timestamp' | 'role' | 'label' | 'snippet'
-> {
-  sourceType: SkillRuntimeEvidencePackSourceType;
-  toolName?: string;
-  isError?: boolean;
-}
-
-export interface SkillRuntimeEvidencePackNode {
-  nodeId: string;
-  nodeKind: 'workflowNode' | 'hardRule';
-  title: string;
-  expectation: string;
-  deterministicStatus: ObservationRuntimeCheckStatus;
-  deterministicReason: string;
-  candidateEvidenceRefs: SkillRuntimeEvidencePackRef[];
-  candidateEvidenceSnippets: string[];
-}
-
-export interface SkillRuntimeEvidencePack {
-  schemaVersion: 1;
-  skillName: string;
-  generatedBy: 'deterministic_rule_pack';
-  definition: {
-    found: boolean;
-    path?: string;
-    truncated?: boolean;
-  };
-  declaredStandards: {
-    hardRules: Array<{ id: string; title: string; expectation: string; source: 'frontmatter' }>;
-    workflowNodes: Array<{ id: string; title: string; expectation: string; workflowId: string; source: 'frontmatter' | 'markdown_headings' }>;
-  };
-  runtimeEvidence: {
-    toolCalls: SkillRuntimeEvidencePackRef[];
-    assistantMessages: SkillRuntimeEvidencePackRef[];
-    userFeedback: SkillRuntimeEvidencePackRef[];
-    artifacts: SkillRuntimeEvidencePackRef[];
-  };
-  nodeEvidence: SkillRuntimeEvidencePackNode[];
-  evidenceQuality: {
-    pollutedSourceCount: number;
-    windowTooNarrow: boolean;
-    missingRuntimeEvidence: boolean;
-    notes: string[];
-  };
-}
-
-export interface ObservationSkillChain {
-  skillName: string;
-  definition: {
-    found: boolean;
-    path?: string;
-    content?: string;
-    truncated?: boolean;
-  };
-  healthCheck: {
-    source: 'doctor-static-rules';
-    hardRules: {
-      declared: boolean;
-      valid: boolean;
-      count: number;
-      rules: SkillHardRule[];
-      errors: string[];
-      advisoryCode?: SkillChainAdvisoryCode;
-    };
-    workflows: {
-      declared: boolean;
-      valid: boolean;
-      branchCount: number;
-      nodeCount: number;
-      workflows: SkillWorkflow[];
-      errors: string[];
-      source: 'frontmatter' | 'markdown_headings' | 'none';
-      advisoryCode?: SkillChainAdvisoryCode;
-    };
-  };
-  runtime: {
-    supported: true;
-    mode: 'deterministic-no-llm';
-    message: string;
-    summary: {
-      invocationCount: number;
-      toolCallCount: number;
-      toolFailureCount: number;
-      passedCount: number;
-      attentionCount: number;
-      manualReviewCount: number;
-    };
-    hardRules: ObservationRuntimeCheck[];
-    workflowNodes: ObservationRuntimeCheck[];
-    evidencePack?: SkillRuntimeEvidencePack;
-  };
 }
