+++ src/observability/inbox.ts
@@ -2,11 +2,12 @@ import { createHash } from 'node:crypto';
 import { existsSync, mkdirSync, readdirSync, readFileSync } from 'node:fs';
 import { join } from 'node:path';
 import { isReportFileName, randomRunToken, reportFilePath } from '../measurement-artifacts/file-names.js';
+import type { ToolCallInfo } from '../types/index.js';
+import type { ObservationExperienceReport } from './contracts/experience.js';
 import type {
   BuildObservationInboxReportOptions,
   ObservationCaptureCoverage,
   ObservationEvidence,
-  ObservationExperienceReport,
   ObservationInboxItem,
   ObservationInboxReport,
   ObservationMessageRef,
@@ -16,10 +17,8 @@ import type {
   ObservationSignalSubtype,
   ObservationSignalType,
   ObservationSkillRollup,
-  ObservationSourceKind,
-  TraceIngestionSummary,
-  ToolCallInfo,
-} from '../types/index.js';
+} from './contracts/inbox.js';
+import type { ObservationSourceKind, TraceIngestionSummary } from './contracts/trace.js';
 import type { GapSignalRef } from '../analysis/contracts.js';
 import { extractGapSignalsFromTrace } from '../analysis/gap-analyzer.js';
 import {
