+++ src/observability/review-state.ts
@@ -12,7 +12,7 @@ import type {
   ObservationReviewStateUpdate,
   ObservationReviewTargetType,
   ObservationReviewVerdict,
-} from '../types/index.js';
+} from './contracts/review.js';
 import { writeJsonFileAtomic } from '../shared/atomic-json.js';
 import { withFileLock } from '../shared/file-lock.js';
 import { normalizeRfc3339Timestamp } from '../shared/timestamp.js';
