+++ src/observability/skill-health-analyzer.ts
@@ -14,7 +14,7 @@
 
 import { buildKnowledgeIndex, computeCoverage, type CoverageReport } from '../analysis/coverage-analyzer.js';
 import { computeGapReport } from '../analysis/gap-analyzer.js';
-import type { TraceIngestionSummary } from '../types/index.js';
+import type { TraceIngestionSummary } from './contracts/trace.js';
 import type { AnalysisEntry, GapReport } from '../analysis/contracts.js';
 import {
   segmentsToAnalysisEntries,
