+++ src/shared/hard-rules.ts
@@ -1,45 +1,12 @@
 import yaml from 'js-yaml';
-
-export interface SkillHardRule {
-  id: string;
-  rule: string;
-  expectedBehavior: string;
-}
-
-export interface SkillWorkflowNode {
-  id: string;
-  action: string;
-}
-
-export interface SkillWorkflow {
-  id: string;
-  description?: string;
-  source?: 'frontmatter' | 'markdown_headings';
-  nodes: SkillWorkflowNode[];
-}
-
-export interface SkillFrontmatterParseResult {
-  hasFrontmatter: boolean;
-  ok: boolean;
-  data?: Record<string, unknown>;
-  error?: string;
-}
-
-export interface SkillHardRulesValidationResult {
-  ok: boolean;
-  hasFrontmatter: boolean;
-  declared: boolean;
-  rules: SkillHardRule[];
-  errors: string[];
-}
-
-export interface SkillWorkflowsValidationResult {
-  ok: boolean;
-  hasFrontmatter: boolean;
-  declared: boolean;
-  workflows: SkillWorkflow[];
-  errors: string[];
-}
+import type {
+  SkillFrontmatterParseResult,
+  SkillHardRule,
+  SkillHardRulesValidationResult,
+  SkillWorkflow,
+  SkillWorkflowNode,
+  SkillWorkflowsValidationResult,
+} from '../skill-definition/contracts.js';
 
 interface YamlErrorLike {
   mark?: { line: number };
