+++ test/architecture/import-boundaries.test.ts
@@ -138,8 +138,11 @@ const RULES: ForbiddenRule[] = [
   {
     from: 'renderer/',
     to: 'observability/',
-    reason: 'renderer 只能通过 facade 访问 observability,不应直接 import observability 内部实现。facade 见 observability/inbox-view-model.ts、observability/feedback-projection.ts、observability/skill-health-analyzer.ts。',
+    reason: 'renderer 只能通过 facade 访问 observability，不应直接 import observability 内部实现。facade 见 observability/view-models/index.ts、observability/inbox-view-model.ts、observability/feedback-projection.ts、observability/skill-health-analyzer.ts。',
     whitelist: [
+      'renderer/conversation-renderer.ts::observability/view-models/index.ts',
+      'renderer/knowledge-debugger-renderer.ts::observability/view-models/index.ts',
+      'renderer/trajectory-evidence.ts::observability/view-models/index.ts',
       // 允许的 facade 访问点(以及它们的 .ts 解析后路径)。
       'renderer/observation-inbox-renderer.ts::observability/inbox-view-model.ts',
       'renderer/observation-inbox-renderer.ts::observability/feedback-projection.ts',
