+++ test/observability/source-record-archive.test.ts
@@ -13,7 +13,7 @@ import {
   writeObservationSourceRecordArchives,
 } from '../../src/observability/source-record-archive.js';
 import { forEachNonEmptyUtf8Line } from '../../src/observability/trace-source.js';
-import type { ObservationSourceRecordArchiveRef } from '../../src/types/index.js';
+import type { ObservationSourceRecordArchiveRef } from '../../src/observability/contracts/inbox.js';
 
 function jsonl(records: unknown[]): string {
   return records.map((record) => JSON.stringify(record)).join('\n');
